Use Option<NonZeroU32> for FormIDs instead of u32
So that Morrowind's lack of FormIDs can be reflected in the type system,
instead of giving its records FormIDs of zero.

This also means that non-Morrowind records with a FormID value of zero
(e.g. TES4 records) will be treated as not having a FormID, which is
more appropriate anyway.