Fix Plugin::overlaps_with() for Morrowind plugins
Morrowind records don't have FormIDs, instead most record types
store an ID in a subrecord and that determine's the record's uniquness.

As a result, it's impossible to tell which records a plugin has are new
and which are overriding a record added by its master(s) without
comparing record IDs with those masters, so
Plugin::count_override_records() continues to return 0 for Morrowind
plugins.

The ID data is hashed for a ~ 25% performance improvement (when fully
parsing Tribunal.esm), but I haven't checked if collisions are likely
to occur within a load order, and will go back to storing the unhashed
data if it's a problem.

I'm not sure that the ID calculation logic is correct, but I based it
off data from Wrye Mash, Elminster and a few other sources.