Use "Light Plugin" instead of "Light Master File"
Whether or not a plugin is light is sometimes independent of whether or
not it is a master, and so the old name was misleading. The new naming
matches what Wrye Bash uses and what the next version of LOOT will use.

Some public function names are affected: the old function names have
kept but deprecated, and renamed equivalents have been added, as it's
not worth the hassle of breaking compatibility over this.