Read strings until first null byte
The MAST subrecords and description part of the HEDR subrecord have
their values treated as null-terminated strings, so anything after
the first null byte is ignored. This has been seen in Skyrim SE's
plugin loading behaviour and in the descriptions displayed for plugins
in Starfield's Creation Kit: the behaviour for other games is untested
but assumed to be the same.

There is also at least one plugin seen "in the wild" that has a MAST
subrecord with extra trailing null bytes:
cheery_AetherArtifactsCraftingAddon_NoDuplicateCrowns.esp from
<https://www.nexusmods.com/skyrimspecialedition/mods/10643>.

As such, ignore the first null byte and everything after it when getting
the description string and vector of master strings for a plugin.