Add Plugin::is_blueprint_plugin()
Starfield plugins have a 0x800 flag value that I'm calling the blueprint flag because one of its effects is that the Creation Kit won't list plugins with that flag set unless bShowBluePrintMastersInDataFiles=1 is set in CreationKit.ini.

The flag has three other effects that I know of. Plugins with the blueprint flag set:

1. Are not visible in Starfield's in-game load order management UI
2. Are removed from Plugins.txt when a save is loaded
3. Load after all plugins of the same scale that do not have the flag set, retaining their load order positions relative to other blueprint plugins.